requirements.txt
python-dotenv
langchain
langchain-community
langchain-experimental
langchain-pinecone
langchain-huggingface
huggingface-hub